#147 - 랭킹 모델 DTO 수정
- 현재 일관 된 제네릭 DTO를 생성해 작품 추천 관련 모델에 적용함
- 추천기반 별모델 수정 및 필요없어진 모델 삭제
- 모델먕 수정
