--- IMAD_Project/Model/Ranking/PopularPosting.swift
@@ -1,14 +1,0 @@
-//
-//  PopularPosting.swift
-//  IMAD_Project
-//
-//  Created by 유영웅 on 3/8/24.
-//
-
-import Foundation
-
-struct PopularPosting:Codable{
-    let status: Int
-    var data: CommunityResponse?
-    let message: String
-}
