--- IMAD_Project/Model/Ranking/Ranking.swift
@@ -1,14 +1,0 @@
-//
-//  Ranking.swift
-//  IMAD_Project
-//
-//  Created by 유영웅 on 1/8/24.
-//
-
-import Foundation
-
-struct Ranking: Codable {
-    let status: Int
-    let data: RankingResponse?
-    let message: String
-}
